+++ terraform/pipeline/ecr.tf
@@ -19,3 +19,18 @@ data "aws_ecr_image" "model_preprocess" {
   repository_name = "fargate/preprocessing"
   image_tag       = terraform.workspace
 }
+
+data "aws_vpc" "default" {
+  default = true
+}
+
+data "aws_subnets" "public" {
+  filter {
+    name   = "vpc-id"
+    values = [data.aws_vpc.default.id]
+  }
+  filter {
+    name   = "map-public-ip-on-launch"
+    values = ["true"]
+  }
+}