+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - New function added to merge the old CQC ratings and the new assessment ratings.
 
 - Polars version of the estimates by job role job and added job to new step function for ind cqc estimates.
+
 - Implemented complex validation for [validate_delta_locations_api_cleaned](projects/_01_ingest/cqc_api/fargate/validate_delta_locations_api_cleaned.py), includes:
   - split into dimensions table with separate validation
   - Pointblank translation of helper functions and new expressions.
@@ -18,6 +19,8 @@ All notable changes to this project will be documented in this file.
 - Function to clean Capacity Tracker care home data by nulling when posts to beds ratio is outside thresholds.
 
 - Model retraining process using Polars and scikit-learn in Fargate
+
+- Tagging added by default to all Terraform elements
 
 ### Changed
 - Migrated Polars validation scripts over to use PointBlank (compatible with >= Python 3.11), so far:
@@ -46,7 +49,13 @@ All notable changes to this project will be documented in this file.
 
 - Removed the deduplication of Capacity Tracker data and used the cleaned Capacity Tracker care home data for imputation.
 
+- Moved the calculation for estimating all Capacity Tracker non-residential filled posts from the diagnostics job to the estimates job.
+
 - Updated the glue script job parameters for SFC-Internal jobs to match SFC-Internal step function.
+
+- Tidy up StepFunctions Terraform code to include a `for_each` to create (most) pipelines dynamically, see the [guide](./terraform/pipeline/step-functions/README.md)
+
+- Explicitly set `required_providers` in `pipeline/main.tf`, in order to version-lock the Terraform AWS provider to the minor version specified
 
 ### Improved